+++ src/App.tsx
@@ -20,9 +20,9 @@ import Withdrawal from 'pages/Setting/Withdrawal';
 import Inquiry from 'pages/Inquiry';
 import Myinquiry from 'pages/Inquiry/Myinquiry';
 import Notice from 'pages/Notice';
-import KakaoLogin from 'KakaoLogin';
-import NaverLogin from 'NaverLogin';
-import GoogleLogin from 'GoogleLogin';
+import KakaoLogin from 'pages/Auth/OAuth/KakaoLogin';
+import NaverLogin from 'pages/Auth/OAuth/NaverLogin';
+import GoogleLogin from 'pages/Auth/OAuth/GoogleLogin';
 
 export default function App(): JSX.Element {
   return (
